1093: Work in progress conmbining add and remove project forms.
